requirements.txt
biopython==1.84
pandas==2.2.2
numpy==1.26.4
scipy==1.13.0
scikit-learn==1.5.0
umap-learn==0.5.6
torch==2.3.0
transformers==4.41.0
py3Dmol==2.1.0
matplotlib==3.9.0
seaborn==0.13.2
plotly==5.22.0
tqdm==4.66.4
pyyaml==6.0.1
h5py==3.11.0
